enh: replace is_infinite_loop with explicit show_loop_count value
This enables users to specify an explicit loop count for playback, instead of being forced to decide between infinite loop or a single show playback.

The previous infinite loop argument, -l, has been retained as the new option value. It now requires an argument which must be >=0 && <= INT_MAX, or may be "i" to specify an infinite loop.

As a hack, if a user knows the length of their show they could use something like cron to execute libreorama with a specific show loop count value such that, when multiplied by the length of the show, libreorama will run for an exact time window without the need for more complex scheduling behavior.